Remove python script, update Pharo section in tex file
In line with the shift to the use of Pharo, the python script (deap_1.py) has been removed. The "Relevant Work/Pharo.tex" has been updated to reflect these changes, particularly the implementation of utilities for the visualization of genetic algorithm in Pharo. Moreover, the updated section includes a more detailed explanation of a genetic algorithm example, with line numbers and additional logging of the algorithm's process for clarity.