Add tcolorbox package and update genetic algorithm details
The tcolorbox package has been added to enhance the display of certain elements like remarks in the thesis document. Text content has been added to the genetic algorithm section to provide detailed explanation about the genetic material and its core properties in the context of genetic algorithms.

The changes also include more detailed information in the Evolutionary Algorithms section under Theoretical_Background, adjustments on the thesis metadata, and image updates on the title page. The font has been updated to improve readability and align with the entire document's look and feel. Several other minor changes to packages and various sections of text were also made.

These changes aim to improve the document’s readability, layout, and content precision. There was a need to elaborate on genetic algorithms and their aspects and to take advantage of the tcolorbox package for an enhanced visual presentation. Design modifications contribute to the document’s overall aesthetics and create a more pleasant reading experience.