Add: Glossary of terms and expanded code highlight configuration
This commit includes several additions and modifications to various files. A glossary of terms used throughout the document has been added in appendix/glossary/_glossary.index.tex, A.tex and M.tex. Definitions are now summarized in a dedicated appendix for ease of reference.

Additional language support and color configurations for the \lstset are also included in preamble/Definitions.tex to improve the appearance of code listings in the document.

The labels in preamble/Definitions.tex have been adjusted to provide more descriptive labeling in the document. In preamble/Theorems.tex, additional configurations for theorem environments have been written, and the fancyhdr page style has been updated with custom header and footer definitions in preamble/config.tex for overall improvements in document layout.

Renamed and reorganized files to improve project structure in .vscode. Worked on improving imperatives for future tense verbs.

The terms are sorted in alphabetical order and represented using the definition environment for clear, consistent formatting. The terms are cross-referenced throughout the document using \vref commands, allowing readers to easily look up unfamiliar terms. The glossary increases the clarity and readability of the document.

The changes to the code highlighting configuration enhance the visual appeal and readability of code blocks within the document. The additions to the fancyhdr setup provide a polished, professional aesthetic. The changes to the folder structure in .vscode simplify navigation within the codebase. Overall, these changes enhance the organization, clarity, and visual appeal of the document and the codebase.