Add LICENSE files, restructure GP section, and update mutation chapters
This commit involves addition of LICENSE files in binary and tex format for clarity on license specifications. The Genetic Programming section in Theoretical_Background has been significantly restructured and updated - the wording modified for cohesion, the extraneous TODO comments eliminated and the bibliography updated. Additionally, mutation chapters for Keen are revised with subsections added for Random Mutator, BitFlip Mutator and the new Swap Mutator. The changes were necessary for clearer, comprehensive explanation of concepts and operators.