Add Inversion Mutator & Update Grow Method
This commit has two major changes: it introduces the Inversion Mutator for Keen library and updates the grow method implementation in the Genetic Programming section under Theoretical_Background.

The Inversion Mutator has been added to improve the functionality and performance of the Keen library in permutation-centric problems. It provides a new way of mutating chromosomes by reversing the sequence of a random subset of genes in a chromosome. Besides providing a detailed implementation and working process of the Inversion Mutator, this commit also improves the clarity and readability of the information provided.

The grow method, an essential random tree generation technique for Genetic Programming, has also been updated. The complex mathematical notations and language have been replaced with a more straightforward Kotlin-like pseudo-code to make it more intuitive to understand. These changes should aid the reader in better understanding the core concept of tree generation in Genetic Programming.
Lastly, some of the sections have been rephrased for greater clarity, ensuring ease of understanding and readability.