Add and detail initial population generation methods in Keen
This commit includes the addition of "Generator.tex" file explaining the generation methods for the initial population in Genetic Programming within the Keen framework. The generation methods, namely 'full' and 'grow', as well as their detailed implementations are explained to aid the understanding of the framework's users. Moreover, references to these methods have been correctly linked in other related files. The aim is to clearly document the initial population generation process, which plays a key role in the GP algorithm.