Add parallelism section, remove redundant content
A section has been added in `_parallelism.index.tex` to discuss how parallelism is achieved in the Keen framework using coroutines. This change highlights the potential for parallelism due to Keen's focus on immutability.

Additionally, redundant user-friendliness content in `Objectives.tex` got removed, and the file `_keen.index.tex` is updated to import the newly created `parallelism` section. These changes provide a more focused and precise understanding of the objectives and structure of the Kotlin-based system.